Split consent outcome from consent responses
